Improved Access for Blind Users

Blind users do not have access to the graphical output from R without printing the content of graph windows to an embosser of some kind. This is not as immediate as is required for efficient access to statistical output. The functions here are created so that blind people can make even better use of R. This includes the text descriptions of graphs, convenience functions to replace the functionality offered in many GUI front ends, and experimental functionality for optimising graphical content to prepare it for embossing as tactile images.

Details

Package:	BrailleR

Type:	Package

Version:	1.2.0

Date:	2025-12-12

License:	GPL-2

Author

A. Jonathan R. Godfrey Volker Sorge, Debra Warren, Timothy P. Bilton and James Thompson with other contributions.

Maintainer: A. Jonathan R. Godfrey <a.j.godfrey@massey.ac.nz>

References

Godfrey, A.J.R. (2013) “Statistical Software from a Blind Person's Perspective: R is the Best, but we can make it better”, The R Journal 5(1), pp73-79.