Send an email based on a template

Template email messages have been created to help you send feedback to the package creator or to join the Blind R Users Group email list. The BlindRUG email list is a forum where we can discuss issues about using R as a blind person. List traffic is light. You can join and leave as you see fit.

JoinBlindRUG()

ThankYou()

Details

You will get an email prepared in your selected email client. If you're wanting to join the BlindRUG list, just press send without editing that message at all. Pretty soon, the server will send you a message that you will need to reply to in order to finalise the subscription. Once you have joined the list you will be sent an initial welcome message.

If you are sending in feedback, then please do tell me how you heard about BrailleR and where you are studying or working.

Author

A. Jonathan R. Godfrey

Value

NULL. These functions are intended for their side effects.