Investigate data objects

Investigate an object, especially useful at any stage in a pipe chain.

CheckIt(x, ...)
check_it(x, ...)

what_is(x, ...)
WhatIs(x, ...)

Arguments

x
The object to be investigated and passed out again.

...
extra parameters to be passed on.

Value

These functions intentionally return (invisibly of course) the original input object so that they can be used within a pipe chain, as is commonly used by the Tidyverse.

Details

The VI() functionality returns a character vector and is not useful inside a pipe chain. In effect, WhatIs() just adds the pipe chain convenience to the VI() tool, while CheckIt() looks at the structure of the data object. The latter is perhaps more useful if you are uncertain that the pipe chain is delivering what you hoped for.

Author

A. Jonathan R. Godfrey

Examples

require(dplyr)

#> Loading required package: dplyr

#> 
#> Attaching package: ‘dplyr’

#> The following objects are masked from ‘package:stats’:
#> 
#>     filter, lag

#> The following objects are masked from ‘package:base’:
#> 
#>     intersect, setdiff, setequal, union

airquality %>% CheckIt() %>% arrange(Ozone) %>% head()

#> 

#> 'data.frame':	153 obs. of  6 variables:
#>  $ Ozone  : int  41 36 12 18 NA 28 23 19 8 NA ...
#>  $ Solar.R: int  190 118 149 313 NA NA 299 99 19 194 ...
#>  $ Wind   : num  7.4 8 12.6 11.5 14.3 14.9 8.6 13.8 20.1 8.6 ...
#>  $ Temp   : int  67 72 74 62 56 66 65 59 61 69 ...
#>  $ Month  : int  5 5 5 5 5 5 5 5 5 5 ...
#>  $ Day    : int  1 2 3 4 5 6 7 8 9 10 ...

#> 

#> 

#>   Ozone Solar.R Wind Temp Month Day
#> 1     1       8  9.7   59     5  21
#> 2     4      25  9.7   61     5  23
#> 3     6      78 18.4   57     5  18
#> 4     7      NA  6.9   74     5  11
#> 5     7      48 14.3   80     7  15
#> 6     7      49 10.3   69     9  24