Unfinished Methods to help vision impaired useRs

A set of methods that will (once coded) extract the most relevant information from a graphical object (or implied set of graphical objects) and display the interpreted results in text form.

The method includes representations of summary methods that are more suitable for blind useRs. For example, the method for a data.frame uses a single line for each variable instead of the normal column layout used by the summary method.

Details

This is the help page for the VI() functions that are not fully functional or below par in some way.

Value

This will vary according to the needs of vision impaired useRs and the specific objects that need to be interpreted.

In general, the output is a series of text strings printed in the console/terminal window in addition to the embedded command's normal functionality.

These functions do not create objects as do many R commands. Manipulations on the objects created by regular R expressions will need those regular expressions issued in addition to those of the VI family of functions.

Author

Jonathan Godfrey