Launches a browser tab to explore SVG diagram

Creates a single HTML file that embeds an SVG diagram and its XML annotations. Then launches a browser tab to allow viewing and interactive exploration of the SVG diagram.

BrowseSVG(file="test", dir=".", key=TRUE, footer = TRUE, view=interactive())

Arguments

file

the filename for the HTML file; this should correspond to basename of an existing SVG and its XML annotations

dir

the folder into which the HTML file is written and where R tries to locate the SVG and XML files

key

include key for explorer's keyboard commands in webpage

footer

Volker, explain please.

view

launch in browser; this is the default when running in an interactive session

Details

An HTML file is written in the given directory and the Javascript library is copied to that location.

Value

NULL. This function exists for its side effects only.

References

to add following first demonstration/publication

Author

Volker Sorge