A convenience function for a two-way analysis

Prepares an analysis of a data set with one response and two predictors that are both factors. An interaction between the two factors is also allowed for. The function creates a number of graphs and tables relevant for the analysis.

TwoFactors(Response, Factor1, Factor2, Inter = FALSE, HSD = TRUE, 
   AlphaE = getOption("BrailleR.SigLevel"), Data = NULL, Filename = NULL, 
    Folder = NULL, VI = getOption("BrailleR.VI"), Latex = getOption("BrailleR.Latex"), 
    View = getOption("BrailleR.View"))

Arguments

Response
Name of the continuous response variable.

Factor1, Factor2
Name the two factors to be included.

Inter
Logical: Should the interaction of the two factors be included?

HSD
Logical: Should Tukey's HSD be evaluated for the data?

AlphaE
The family-wise Type I error rate for Tukey's HSD calculations.

Data
Name the data.frame that includes the three variables of interest.

Filename
Name of the Rmarkdown and HTML files to be created. A default will be created that uses the names of the variables if this is left set to NULL.

Folder
Name of the folder to store graph and LaTeX files. A default will be created based on the name of the data.frame being used.

VI
Logical: Should the VI method for blind users be employed?

Latex
Logical: Should the tabulated sections be saved in LaTeX format?

View
Logical: Should the HTML file be opened for inspection?

Details

to complete

Value

This function is used for creation of the files saved in the working directory and a few of its subdirectories.

Author

Timothy P. Bilton and A. Jonathan R. Godfrey

See also

Examples

DIR = getwd()
setwd(tempdir())
TG <- ToothGrowth
TG$dose <- as.factor(TG$dose)

# Without interaction
TwoFactors('len','supp','dose',Data=TG, Inter=FALSE)

#> Error in get(DataName): object 'TG' not found

# With two-way interaction
TwoFactors('len', 'supp', 'dose', Data=TG, Inter=TRUE)

#> Error in get(DataName): object 'TG' not found

# For unbalanced data
TG_Unb <- TG[-c(53:60),]
TwoFactors('len', 'supp', 'dose', Data=TG_Unb, Inter=TRUE)

#> Error in get(DataName): object 'TG_Unb' not found

rm(TG); rm(TG_Unb)
# N.B. Various files and a folder were created in a temporary directory. 
# Please investigate them to see how this function worked.
setwd(DIR)