Count points in a scatter plot

count the number of points that fall into various sized subparts of a scatter plot. The graphing region can be split into cells based on a uniform or normal marginal distribution separately for the x and y variables.

WhereXY(x, y = NULL, grid = c(4, 4), xDist = "uniform", 
       yDist = xDist, addmargins=TRUE)

Arguments

x,y

vectors of x coordinates. If y is not specified, the function expects x to be a two-column matrix with x and y values in columns 1 and 2 respectively.

grid

pair of values to specify the way the graph is to be split into parts. Specify x and then y.

xDist,yDist

the distribution the variables might be expected to follow. The default is to consider uniformly distributed but any alternative text will lead to an assumption of both margins being normally distributed.

addmargins

logical: should sums be added to both rows and columns.

Value

A text description of the number of points in each subregion of the scatter plot. The table of counts can then be compared to the expected number of points in each subregion.

Author

A. Jonathan R. Godfrey

Examples

x=rnorm(50)
y=rnorm(50)
WhereXY(x,y)
#>      1  2  3 4 Sum
#> 4    4  2  0 0   6
#> 3    6  4  7 3  20
#> 2    3  7  8 1  19
#> 1    1  2  2 0   5
#> Sum 14 15 17 4  50
WhereXY(x,y, c(3,4))
#>      1  2 3 Sum
#> 4    4  2 0   6
#> 3    7  8 5  20
#> 2    8  9 2  19
#> 1    1  3 1   5
#> Sum 20 22 8  50
WhereXY(x,y, xDist="other")
#>      1  2  3  4 Sum
#> 4    4  2  2  1   9
#> 3    6  1  3  7  17
#> 2    3  5  1  5  14
#> 1    1  3  5  1  10
#> Sum 14 11 11 14  50